+++ ServerData.cs
@@ -89,6 +89,8 @@ public static class RouteRegistry
         new("/routes", "GET"),
         new("/login", "POST"),
         new("/logout", "POST"),
+        new("/player/data", "GET"),
+        new("/player/data", "PATCH"),
     };
 }
 
@@ -96,5 +98,5 @@ public record RouteInfo(string Route, string Method);
 
 public static class APIVersion
 {
-    public const string Version = "1.4.0";
+    public const string Version = "1.5.0";
 }